[serve][llm] 4/4: Add benchmark docs and refactor into submodules
- Add README with methodology, usage docs, and interactive command reference
- Extract multiturn_bench.py (1900 lines) into focused submodules:
  models.py, http_client.py, text_gen.py, metrics.py, turn.py,
  reporting.py, runners.py
- Delete multiturn_bench.py (all consumers migrated to submodules)
- Simplify WorkloadSpec solver: unified formula with no special cases
- Add actionable feasibility suggestions to solver error messages
- Fix WorkloadSpec.shared_system_prompt_ratio default to 1.0
